requirements.txt
# ============================================
# FB Scrape Ideas - Abstract Dependencies
# ============================================
# This file defines the top-level dependencies.
# For deterministic builds and CI/CD, use: requirements.lock
# ============================================
# ============================================
# FB Scrape Ideas - Abstract Dependencies
# ============================================
# This file defines the top-level dependencies.
# For deterministic builds and CI/CD, use: requirements.lock
# ============================================
# HTTP & Web Scraping
requests>=2.32.5
beautifulsoup4>=4.14.3
lxml>=6.0.2

# Browser Automation
selenium>=4.39.0
webdriver-manager>=4.0.2
playwright>=1.50.0


# Google Gemini AI (new SDK - google-genai replaces deprecated google-generativeai)
google-genai>=1.0.0

# OpenAI-Compatible API (also works with Ollama, LM Studio, OpenRouter, etc.)
openai>=2.14.0

# Environment & Configuration
python-dotenv>=1.2.1
cryptography>=44.0.0


# Date/Time Parsing
python-dateutil>=2.9.0.post0
dateparser>=1.2.2

# Utilities
tenacity>=9.1.2
